Move unified messenger to its own package, and explode classes to their own class files so that they may then be marked public
