+++ src/games/strategy/engine/framework/startup/mc/ServerModel.java
@@ -52,7 +52,7 @@ import games.strategy.engine.message.IChannelMessenger;
 import games.strategy.engine.message.IRemoteMessenger;
 import games.strategy.engine.message.RemoteMessenger;
 import games.strategy.engine.message.RemoteName;
-import games.strategy.engine.message.UnifiedMessenger;
+import games.strategy.engine.message.unifiedmessenger.UnifiedMessenger;
 import games.strategy.net.IConnectionChangeListener;
 import games.strategy.net.IMessenger;
 import games.strategy.net.IMessengerErrorListener;
