+++ src/games/strategy/engine/message/RemoteMethodCallResults.java
@@ -10,7 +10,7 @@ import java.io.ObjectOutput;
  * Note that either one of m_rVal or m_exception will be null,
  * since the method can either throw or return
  */
-class RemoteMethodCallResults implements Externalizable {
+public class RemoteMethodCallResults implements Externalizable {
   private static final long serialVersionUID = 4562274411264858613L;
   private Object m_rVal;
   // throwable implements Serializable
