+++ src/games/strategy/engine/message/SpokeInvocationResults.java
@@ -1,5 +1,6 @@
 package games.strategy.engine.message;
 
+import games.strategy.engine.message.unifiedmessenger.InvocationResults;
 import games.strategy.net.GUID;
 
 public class SpokeInvocationResults extends InvocationResults {
