+++ src/games/strategy/engine/message/SpokeInvoke.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.io.ObjectInput;
 import java.io.ObjectOutput;
 
+import games.strategy.engine.message.unifiedmessenger.Invoke;
 import games.strategy.net.GUID;
 import games.strategy.net.INode;
 import games.strategy.net.Node;
