+++ src/games/strategy/net/Messengers.java
@@ -4,7 +4,7 @@ import games.strategy.engine.message.ChannelMessenger;
 import games.strategy.engine.message.IChannelMessenger;
 import games.strategy.engine.message.IRemoteMessenger;
 import games.strategy.engine.message.RemoteMessenger;
-import games.strategy.engine.message.UnifiedMessenger;
+import games.strategy.engine.message.unifiedmessenger.UnifiedMessenger;
 
 /**
  * Convenience grouping of a messenger, remote messenger and channel messenger.
