+++ test/games/strategy/engine/message/unifiedmessenger/ChannelMessengerTest.java
@@ -1,10 +1,14 @@
-package games.strategy.engine.message;
+package games.strategy.engine.message.unifiedmessenger;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import java.io.IOException;
 
+import games.strategy.engine.message.ChannelMessenger;
+import games.strategy.engine.message.IChannelSubscribor;
+import games.strategy.engine.message.RemoteName;
+import games.strategy.engine.message.UnifiedMessengerHub;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
